# web/src/api-docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateVenueRequest" : {
        "properties" : {
          "accessibility" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "addressLine" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "capacity" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "city" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "indoorOutdoor" : {
            "enum" : [ "INDOOR", "OUTDOOR", "MIXED" ],
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          },
          "mapUrl" : {
            "maxLength" : 2048,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 160,
            "minLength" : 0,
            "type" : "string"
          },
          "notes" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "parking" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "photoPath" : {
            "maxLength" : 512,
            "minLength" : 0,
            "pattern" : "venue-images/[A-Za-z0-9._-]+",
            "type" : "string"
          }
        },
        "required" : [ "addressLine", "name" ],
        "type" : "object"
      },
...
      "EmailCodeRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
...
      "PageResponseVenueResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/VenueResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UpdateVenueRequest" : {
        "properties" : {
          "accessibility" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "addressLine" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "capacity" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "city" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "indoorOutdoor" : {
            "enum" : [ "INDOOR", "OUTDOOR", "MIXED" ],
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "maximum" : 90.0,
            "minimum" : -90.0,
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "maximum" : 180.0,
            "minimum" : -180.0,
            "type" : "number"
          },
          "mapUrl" : {
            "maxLength" : 2048,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 160,
            "minLength" : 0,
            "type" : "string"
          },
          "notes" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "parking" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "photoPath" : {
            "maxLength" : 512,
            "minLength" : 0,
            "pattern" : "venue-images/[A-Za-z0-9._-]+",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "VenueResponse" : {
        "properties" : {
          "accessibility" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "addressLine" : {
            "type" : "string"
          },
          "capacity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "city" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "indoorOutdoor" : {
            "enum" : [ "INDOOR", "OUTDOOR", "MIXED" ],
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "mapUrl" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "parking" : {
            "type" : "string"
          },
          "photoPath" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "REST API for the TeamMarhaba backend service.",
    "title" : "TeamMarhaba Backend API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/venues" : {
      "get" : {
        "operationId" : "list",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "active",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseVenueResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      },
      "post" : {
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateVenueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VenueResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      }
    },
    "/api/v1/admin/venues/{id}" : {
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VenueResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      },
      "patch" : {
        "operationId" : "update",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateVenueRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VenueResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      }
    },
    "/api/v1/admin/venues/{id}/deactivate" : {
      "post" : {
        "operationId" : "deactivate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VenueResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      }
    },
    "/api/v1/admin/venues/{id}/reactivate" : {
      "post" : {
        "operationId" : "reactivate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VenueResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "venue-admin-controller" ]
      }
    },
...
    "/api/v1/auth/email-code/request" : {
      "post" : {
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EmailCodeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "email-code-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}